Conflicts:
	lib/action.php
	plugins/OStatus/actions/ostatusinit.php

+++ actions/deletenotice.php
@@ -32,6 +32,7 @@ if (!defined('STATUSNET') && !defined('LACONICA')) {
     exit(1);
 }
 
+// @todo FIXME: documentation needed.
 class DeletenoticeAction extends Action
 {
     var $error        = null;
@@ -45,10 +46,18 @@ class DeletenoticeAction extends Action
         parent::prepare($args);
 
         $this->user   = common_current_user();
+
+        if (!$this->user) {
+            // TRANS: Error message displayed trying to delete a notice while not logged in.
+            common_user_error(_('Not logged in.'));
+            exit;
+        }
+
         $notice_id    = $this->trimmed('notice');
         $this->notice = Notice::staticGet($notice_id);
 
         if (!$this->notice) {
+            // TRANS: Error message displayed trying to delete a non-existing notice.
             common_user_error(_('No such notice.'));
             exit;
         }
@@ -63,11 +72,9 @@ class DeletenoticeAction extends Action
     {
         parent::handle($args);
 
-        if (!common_logged_in()) {
-            common_user_error(_('Not logged in.'));
-            exit;
-        } else if ($this->notice->profile_id != $this->user_profile->id &&
+        if ($this->notice->profile_id != $this->user_profile->id &&
                    !$this->user->hasRight(Right::DELETEOTHERSNOTICE)) {
+            // TRANS: Error message displayed trying to delete a notice that was not made by the current user.
             common_user_error(_('Can\'t delete this notice.'));
             exit;
         }
@@ -87,7 +94,6 @@ class DeletenoticeAction extends Action
      *
      * @return void
      */
-
     function showPageNotice()
     {
         $instr  = $this->getInstructions();
@@ -100,12 +106,14 @@ class DeletenoticeAction extends Action
 
     function getInstructions()
     {
+        // TRANS: Instructions for deleting a notice.
         return _('You are about to permanently delete a notice. ' .
                  'Once this is done, it cannot be undone.');
     }
 
     function title()
     {
+        // TRANS: Page title when deleting a notice.
         return _('Delete notice');
     }
 
@@ -118,7 +126,6 @@ class DeletenoticeAction extends Action
      *
      * @return void
      */
-
     function showForm($error = null)
     {
         $this->error = $error;
@@ -130,7 +137,6 @@ class DeletenoticeAction extends Action
      *
      * @return void
      */
-
     function showContent()
     {
         $this->elementStart('form', array('id' => 'form_notice_delete',
@@ -138,9 +144,11 @@ class DeletenoticeAction extends Action
                                           'method' => 'post',
                                           'action' => common_local_url('deletenotice')));
         $this->elementStart('fieldset');
+        // TRANS: Fieldset legend for the delete notice form.
         $this->element('legend', null, _('Delete notice'));
         $this->hidden('token', common_session_token());
         $this->hidden('notice', $this->trimmed('notice'));
+        // TRANS: Message for the delete notice form.
         $this->element('p', null, _('Are you sure you want to delete this notice?'));
         $this->submit('form_action-no',
                       // TRANS: Button label on the delete notice form.
@@ -172,7 +180,10 @@ class DeletenoticeAction extends Action
         }
 
         if ($this->arg('yes')) {
-            $this->notice->delete();
+            if (Event::handle('StartDeleteOwnNotice', array($this->user, $this->notice))) {
+                $this->notice->delete();
+                Event::handle('EndDeleteOwnNotice', array($this->user, $this->notice));
+            }
         }
 
         $url = common_get_returnto();
